# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend API for a multi-tenant (SaaS) ophthalmology clinic\nmanagement system: patients, visits, billing/POS, inventory,\nlab jobs, reports, and optical calculators.\n\nEvery route under /api (other than /api/auth/*, /api/signup,\nand the /api/auth/password-reset|email endpoints) requires a\nBearer access token, and most additionally require the\ncaller's clinic to have an active subscription.",
        "title": "Opthalmic Management System API",
        "contact": {
            "name": "API Support"
        },
        "license": {
            "name": "Proprietary"
        },
        "version": "1.0"
    },
    "host": "localhost:8082",
    "basePath": "/api",
    "paths": {
...
        "/admin/clinic": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Admin-only.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "clinic"
                ],
                "summary": "List clinics",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/http.response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/dto.ClinicResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Admin-only.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "clinic"
                ],
                "summary": "Create clinic settings",
                "parameters": [
                    {
                        "description": "Clinic details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.ClinicRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/http.response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.ClinicResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/admin/clinic/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Admin-only.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "clinic"
                ],
                "summary": "Get clinic settings",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Clinic ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/http.response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.ClinicResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/http.errorResponse"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Admin-only.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "clinic"
                ],
                "summary": "Update clinic settings",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Clinic ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Fields to update",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.ClinicRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/http.response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.ClinicRequest": {
            "type": "object",
            "required": [
                "clinic_name",
                "registration_no"
            ],
            "properties": {
                "address": {
                    "type": "string"
                },
                "clinic_name": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "registration_no": {
                    "type": "string"
                },
                "report_footer": {
                    "type": "string"
                },
                "tagline": {
                    "type": "string"
                }
            }
        },
        "dto.ClinicResponse": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "clinic_name": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "registration_no": {
                    "type": "string"
                },
                "report_footer": {
                    "type": "string"
                },
                "tagline": {
                    "type": "string"
                },
                "updated_by": {
                    "type": "string"
                }
            }
        },
...
        "http.errorResponse": {
            "type": "object",
            "properties": {
                "messages": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "Error message 1",
                        " Error message 2"
                    ]
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "http.response": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string",
                    "example": "Success"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and the access token, e.g. \"Bearer eyJhbGci...\".",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}